+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to [bpmn-js](https://github.com/bpmn-io/bpmn-js) are documen
 ___Note:__ Yet to be released changes appear here._
 
 * `FEAT`: add initial snapping when creating associations
+* `FIX`: allow message flows between collapsed pools
 
 ## 1.1.1